python-tracker/requirements.txt
mss>=9.0.0
opencv-python>=4.8.0
requests>=2.31.0
pystray>=0.19.0
Pillow>=10.0.0
pynput>=1.7.6
pyinstaller>=6.0.0
